sove bug of load xml

+++ src/main/java/com/oop_paint/Shapes/Circle.java
@@ -18,7 +18,7 @@ public class Circle extends Shape {
         return radius;
     }
 
-    public void setRadius(int radius) {
+    public void setRadius(double radius) {
         this.radius = radius;
     }
